+++ src/main/java/Reika/DragonAPI/Instantiable/Data/Immutable/DecimalPosition.java
@@ -199,7 +199,7 @@ public final class DecimalPosition {
 
 	public void dropItem(World world, ItemStack is, double vscale) {
 		if (world != null && !world.isRemote) {
-			ReikaItemHelper.dropItem(world, xCoord+rand.nextDouble(), yCoord+rand.nextDouble(), zCoord+rand.nextDouble(), is, vscale);
+			ReikaItemHelper.dropItem(world, xCoord+rand.nextFloat(), yCoord+rand.nextFloat(), zCoord+rand.nextFloat(), is, vscale);
 		}
 	}
