+++ src/main/java/Reika/DragonAPI/Libraries/ReikaEntityHelper.java
@@ -633,7 +633,7 @@ public final class ReikaEntityHelper extends DragonAPICore {
 	/** Spawns a bunch of particles around an entity. Args: Particle Type, Entity, number of particles */
 	public static void spawnParticlesAround(String part, Entity e, int num) {
 		for (int k = 0; k < num; k++)
-			e.worldObj.spawnParticle(part, e.posX-0.6+1.2*rand.nextDouble(), e.posY+e.height/2-0.6+1.2*rand.nextDouble(), e.posZ-0.6+1.2*rand.nextDouble(), -0.2+0.4*rand.nextDouble(), 0.4*rand.nextDouble(), -0.2+0.4*rand.nextDouble());
+			e.worldObj.spawnParticle(part, e.posX-0.6+1.2*rand.nextFloat(), e.posY+e.height/2-0.6+1.2*rand.nextFloat(), e.posZ-0.6+1.2*rand.nextFloat(), -0.2+0.4*rand.nextFloat(), 0.4*rand.nextFloat(), -0.2+0.4*rand.nextFloat());
 	}
 
 	/** Returns the Entity ID from entity class. Args: Entity Class */
